Add Node status model and migrations
